Move Scout related classes into `\Nuwave\Lighthouse\Scout` (#1698)
