+++ CHANGELOG.md
@@ -18,6 +18,7 @@ You can find and compare releases at the [GitHub release page](https://github.co
 ### Changed
 
 - Clarify semantics of combining `@search` with other directives https://github.com/nuwave/lighthouse/pull/1691
+- Move Scout related classes into `\Nuwave\Lighthouse\Scout` https://github.com/nuwave/lighthouse/pull/1698
 
 ## 5.1.0
